refactor: remove redundant empty check
You don't need to check for an empty array before your 'for-each' loop;
when the array is empty the loop doesn't have any elements to loop over
in the first place and basically does nothing.
So validating the arrays before the loops doesn't change the outcome of
your functions but improves code simplicity.